set free standing functions in murmurhash to static
If one includes the bloom filter header in multiple compilation units
and tries to link them all together at the end, the linker will
complain about redefinitions of some functions in the detail namespace.
These functions are only used internally and are not part of the
external API. Setting them to static gets rid of the linker error.